+++ ghostos/libraries/terminal/__init__.py
@@ -1,0 +1,3 @@
+from ghostos.libraries.terminal.abcd import Terminal, TerminalContext
+
+from ghostos.libraries.terminal.ubuntu_terminal import UbuntuTerminalProvider